requirements.txt
anthropic>=0.78.0
pyyaml>=6.0
requests>=2.31.0
python-dotenv>=1.0.0
gradio>=4.0.0